refactor: combine init method into service
