+++ src/components/Vuetify/index.ts
@@ -1,6 +1,5 @@
 import application from './mixins/application'
-import theme from './mixins/theme'
-import ThemeService from './mixins/app-theme'
+import ThemeService, { initTheme } from './services/ThemeService'
 import icons from './mixins/icons'
 import options from './mixins/options'
 import genLang from './mixins/lang'
@@ -31,7 +30,7 @@ const Vuetify: VuetifyPlugin = {
         lang,
         options: options(opts.options),
         rtl: opts.rtl,
-        theme: theme(opts.theme)
+        theme: initTheme(opts.theme)
       },
       methods: {
         goTo,
